Moved S3Url util class to util.py
