+++ tests/util.py
@@ -1,4 +1,6 @@
 import os
+import boto3
+from urllib.parse import urlparse
 from dbt.config.project import PartialProject
 
 
@@ -110,3 +112,38 @@ def get_s3_location():
 def get_role_arn():
     return os.environ.get("DBT_GLUE_ROLE_ARN", f"arn:aws:iam::{get_account_id()}:role/GlueInteractiveSessionRole")
 
+
+def cleanup_s3_location(path, region):
+    client = boto3.client("s3", region_name=region)
+    S3Url(path).delete_all_keys_v2(client)
+
+
+class S3Url(object):
+    def __init__(self, url):
+        self._parsed = urlparse(url, allow_fragments=False)
+
+    @property
+    def bucket(self):
+        return self._parsed.netloc
+
+    @property
+    def key(self):
+        if self._parsed.query:
+            return self._parsed.path.lstrip("/") + "?" + self._parsed.query
+        else:
+            return self._parsed.path.lstrip("/")
+
+    @property
+    def url(self):
+        return self._parsed.geturl()
+
+    def delete_all_keys_v2(self, client):
+        bucket = self.bucket
+        prefix = self.key
+
+        for response in client.get_paginator('list_objects_v2').paginate(Bucket=bucket, Prefix=prefix):
+            if 'Contents' not in response:
+                continue
+            for content in response['Contents']:
+                print("Deleting: s3://" + bucket + "/" + content['Key'])
+                client.delete_object(Bucket=bucket, Key=content['Key'])
